Update enable_l3 param to also accept booleans
The enable_l3 param used to simply pass the recieved value
to the template for output, so they had to be the string
'yes' or 'no', as expected by ODL. The template now has
logic for accepting true or 'yes' and writing 'yes' or
false or 'no' and writing 'no.

The rspec-puppet tests, Beaker tests and docs are also updated.

Relevant to #87

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>