Remove 14.04 Beaker tests from acpt rake task
Remove the Ubuntu 14.04 Docker-based Beaker tests from the rake task
called "acceptance". This is more for clarity than because anything
changed. Ubuntu support is not-well-supported, and tarball-based
installs are less-than-ideal and (hopefully) soon to be replaced by .deb
installs (once ODL Int/Pack has a .deb). We wouldn't call a release of
puppet-opendaylight "failing" if it was failing the Ubuntu Beaker tests,
so we shouldn't include them in this rake task.

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>