Adds ability to pass Java options to ODL startup
Previously we always set the Java options to a single value which will
only run Java with ipv4 stack.  There is a requirement to be able to set
other parameters like heap seize or enable garbage collection.  This
change exposes a parameter where many Java options can be set via a
string.

Signed-off-by: Tim Rozet <tdrozet@gmail.com>